reuse html escape query params function
