+++ lib/core.js
@@ -12,7 +12,7 @@
     import * as htmlUtils from './html-utils.js';
     import * as metaUtils from './plugins/system/meta/utils.js';
     import mediaPlugin from './plugins/validators/media.js';
-    import { difference, intersection } from '../utils.js';
+    import { difference, intersection, normalizeQueryOptionValue } from '../utils.js';
 
     const plugins = pluginLoader._plugins,
         pluginsModules = pluginLoader._pluginsModules,
@@ -1509,21 +1509,6 @@
         }
     }
 
-    function normalizeValue(value) {
-        if (value === 'true') {
-            return true;
-        }
-        if (value === 'false') {
-            return false;
-        }
-        if (value.match && value.match(/^\d+$/)) {
-            return parseInt(value);
-        }
-        if (value.match && value.match(/^(\d+)?\.\d+$/)) {
-            return parseFloat(value);
-        }
-        return value;
-    }    
 
     function getQueryOptionsFromEntries(queryEntries) {
         const _RE = /^_.+/;
@@ -1537,7 +1522,7 @@
 
         queryEntries.forEach((value, key) => {
             if (key.length > 1 && _RE.test(key)) {
-                value = normalizeValue(value);
+                value = normalizeQueryOptionValue(value);
                 var realKey = key.substr(1);
                 result[realKey] = value;
             }
